requirements.txt
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
h11==0.14.0
idna==3.6
jsonpickle==3.0.2
lxml==5.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.3.0.post0
pycparser==2.21
Pygments==2.17.2
PySocks==1.7.1
PyYAML==6.0.1
rich==13.7.0
selenium==4.17.2
shellingham==1.5.4
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
